requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
filelock==3.13.1
flatbuffers==25.2.10
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
humanfriendly==10.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.3
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.44.0
MarkupSafe==2.1.5
mpmath==1.3.0
narwhals==1.32.0
networkx==3.3
ninja==1.11.1.4
numba==0.61.0
numpy==2.1.2
onnxruntime-gpu==1.21.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.7
pooch==1.8.2
protobuf==5.29.4
pyarrow==19.0.1
pydeck==0.9.1
PyMatting==1.1.13
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
rembg==2.0.65
requests==2.32.3
rpds-py==0.24.0
scikit-image==0.25.2
scipy==1.15.2
six==1.17.0
smmap==5.0.2
streamlit==1.44.0
sympy==1.13.1
tenacity==9.0.0
tifffile==2025.3.30
toml
torch==2.5.1
torchaudio==2.5.1
torchgeometry==0.1.2
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
typing_extensions
tzdata==2025.2
urllib3==2.3.0
watchdog==6.0.0
inplace-abn
